stuff... 3.5 hours
So.
What I did was basically finishing ALL the UMLs.
Don't judge... it actually took me 3.5 hours since we didn't really do
it for the last 2 submissions plus we did the class diagram in the wrong
way so ya...
And she had extra requirements for the sequence so...
well, finished